+++ odoo/songs/install/data_demo.py
@@ -130,6 +130,17 @@ def import_sales_order(ctx):
 
 
 @anthem.log
+def import_purchase_order(ctx):
+    """ Importing purchases order from csv """
+    model = ctx.env['purchase.order'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 'data/demo/purchase_order_head.csv', model)
+    load_csv(ctx, 'data/demo/purchase_order_line.csv',
+             'purchase.order.line')
+
+
+@anthem.log
 def import_waves(ctx):
     """ Importing waves from csv """
     load_csv(ctx, 'data/demo/wave.csv', 'stock.picking.wave')
@@ -183,5 +194,6 @@ def main(ctx):
     import_bank(ctx)
     import_bank_account(ctx)
     import_sales_order(ctx)
+    import_purchase_order(ctx)
     import_invoices_supplier(ctx)
     return