+++ odoo/songs/install/data_full.py
@@ -168,6 +168,18 @@ def import_sales_order(ctx):
 
 
 @anthem.log
+def import_purchase_order(ctx):
+    """ Importing purchases order from csv """
+    model = ctx.env['purchase.order'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/purchase_order_head.csv',
+             model)
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/purchase_order_line.csv',
+             'purchase.order.line')
+
+
+@anthem.log
 def import_bom(ctx):
     """ Importing bill of materials from csv """
     load_csv(ctx, 's3://prod-sf-odoo-data/install/mrp_bom.csv', 'mrp.bom')
@@ -235,6 +247,7 @@ def main(ctx):
     import_bank(ctx)
     import_bank_account(ctx)
     import_sales_order(ctx)
+    import_purchase_order(ctx)
     import_bom(ctx)
     import_invoices_supplier(ctx)
     return